added footer icons and reorganised the files imports
