+++ Project-I-Social-Card/social-card/src/App.js
@@ -1,41 +1,8 @@
 import React from 'react';
 import './App.css';
 
-import './components/HeaderComponents/HeaderContent.js'
 
-// starting with the smallest piece
-// header title
-function HeaderTitle ({titleText}){
-  return (
-    <h1>{titleText} <span>@LambdaSchool . 27 may</span></h1>
-  );
-}
-// The header Thumbnail
-function HeaderThumb ({link}) {
-  return (
-    <img src={link} className="logo" alt=""/>
-  );
-}
-
-// The content
-function HeaderContent ({headerText}) {
-  return (
-    <div>
-      <HeaderTitle titleText ="Lambda School"/>
-      <p> {headerText} </p>
-    </div>
-  );
-}
-function BuildHeader () {
-  return (
-    <div className = "headerContainer">
-      <HeaderThumb link="https://tk-assets.lambdaschool.com/1c1b7262-cf23-4a9f-90b6-da0d3c74a5c6_lambdacrest.png" />
-      <HeaderContent headerText="Let's learn react by builing simple interfaces components.
-      Don't try to overthink it, just keep it simple abd have fun. Once you feel comfortable using
-      components you are well on your way to master React !"/>     
-    </div>
-  );
-}
+import BuildHeader from './components/HeaderComponents/HeaderContainer'
 
 // the card section
 
@@ -70,11 +37,26 @@ function CardBuilder (){
   );
 }
 
+// Footer start here
+
+function FooterBuilder () {
+  return (
+    <footer>
+      <div className="icon-countainer">
+        <i class="far fa-comment"></i>
+        <i class="fas fa-retweet"></i>
+        <i class="far fa-heart"></i>
+      </div>
+    </footer>
+  );
+}
+
 const App = () => {
   return (
     <div className="mainContainer">
       <BuildHeader />
       <CardBuilder />
+      <FooterBuilder />
     </div>
   );
 };
