+++ Project-I-Social-Card/social-card/src/components/HeaderComponents/Header.css
@@ -9,9 +9,9 @@ span {
 }
 
 .logo {
-    width: 10%;
-    height: 10%;
-    padding-right: 2rem;
+    width: 5%;
+    height: 5%;
+    padding: .5rem;
 }
 
 .headerContainer{
