+++ Project-I-Social-Card/social-card/src/components/HeaderComponents/HeaderContent.js
@@ -2,10 +2,21 @@ import React from 'react';
 import './Header.css';
 
 // The content
-function HeaderContent ({headerText}) {
-    return (
-      <p> {headerText} </p>
-    );
-  }
 
-  export default HeaderContent;
+// header title
+function HeaderTitle ({titleText}){
+  return (
+    <h1>{titleText} <span>@LambdaSchool . 27 may</span></h1>
+  );
+}
+
+function HeaderContent ({headerText}) {
+  return (
+    <div>
+      <HeaderTitle titleText ="Lambda School"/>
+      <p> {headerText} </p>
+    </div>
+  );
+}
+
+export default HeaderContent;
