+++ Project-I-Social-Card/social-card/src/components/HeaderComponents/ImageThumbnail.js
@@ -1,2 +1,10 @@
 import React from 'react';
 import './Header.css';
+
+function HeaderThumb ({link}) {
+    return (
+      <img src={link} className="logo" alt=""/>
+    );
+  }
+
+  export default HeaderThumb;
